requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
yfinance>=0.1.70
streamlit>=1.12.0
plotly>=5.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
textblob>=0.15.3
newspaper3k>=0.2.8
tensorflow>=2.8.0
statsmodels>=0.13.0
seaborn>=0.11.0
pandas-datareader>=0.10.0
requests>=2.25.0
